Try to use symbolic link for common DB test
